SubTasks are the new feature
I have worked on adding the subtask feature for all user accounts. Each individual task can have an infinite number of subtasks. Those subtasks are made to lead to their parent task. Each subtask has a parent task and an identification id which is the primary key of the parent task.
Hope this feature is handy!